Cursor Images 1.0
Here is my first batch of images :DD
These are just cursor images for buttons and visuals (i.e. upgrades and tower/s).
And now I shall explain how the names of images work (which is how it will always work when I upload files).
The names of files basically brake up into categories with the words on the left being more broad and ones on the right being more detailed.
In this case, the first part tells what function/object it is (i.e. cursor)
The second part tells what I used to edit the picture with (i.e. Overlay. Original means it's the image used to create all other images in it's category). I suggest that when using these images, seperate them using the second category in the file name (i.e. cursorOverlay will be used for general upgrades, and cursorTemp and cursorSpotlight used for VERY specific upgrades that usually only occur once).
The third part tells what version it is. This may range from an obvious number (such as 1, 2, 3, or 1.0, 1.1, 1.2), or to specific details (i.e. curs